// src/app/shared/enums/beverages-characteristics-enum.ts
export enum CoffeePreparationEnum {
  WITH_GROUNDS = 'fusiasta',
  COFFEE_MAKER = 'z ekspresu',
  INSTANT = 'rozpuszczalna'
}

export enum CoffeeStrenghtEnum {
  WATERY = 'lura',
  STANDARD = 'normalna',
  STRONG = 'szatan'
}

export enum MilkEnum {
  NONE = 'brak',
  ONE_THIRD = '1/3 kubka',
  TWO_THIRDS = '2/3 kubka'
}

export enum SugarEnum {
  NONE = 'brak',
  ONE_SPOON = '1 łyżeczka',
  TWO_SPOONS = '2 łyżeczki',
  THREE_SPOONS = '3 łyżeczki'
}

export enum TeaTypeEnum {
  GREEN = 'zielona',
  BLACK = 'czarna',
  FRUITY = 'owocowa'
}